% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_by_agem.R
\name{cases_by_agem}
\alias{cases_by_agem}
\title{This function counts VBD cases per agem.}
\usage{
cases_by_agem(path_vbd, vbd, status_case, country, cve_edo = NULL)
}
\arguments{
\item{path_vbd}{Is the directory where the file is located.}

\item{vbd}{Is the parameter for define the vector-borne diseases}

\item{status_case}{It is a numerical value to select the Status of the case. There are three options, 1 for the probable cases database, 2 for the confirmed cases database, and 3 for the discarded cases dataset.}

\item{country}{It is a logical value to define if the result is by state (set TRUE) or by country (set FALSE)}

\item{cve_edo}{Is a string of the state id-}
}
\value{
a sf object
}
\description{
This function counts VBD cases per agem.
}
\examples{
1+1
}
