% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geocode.R
\name{data_geocode}
\alias{data_geocode}
\title{data_geocode}
\usage{
data_geocode(infile, data)
}
\arguments{
\item{infile}{is the name of file create with \link[denhotspots]{subset_den}.}

\item{data}{is a string for define the addreses or data. if data TRUE, is return the dataset.}
}
\value{
a vector.
}
\description{
this function creates an address vector.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
