% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densin-package.R
\docType{package}
\name{densin-package}
\alias{densin}
\alias{densin-package}
\title{densin: densin is a package designed to perform exploratory analysis of the SINAVE dengue dataset}
\description{
The package reads the SINAVE Vector-Borne Diseases dataset, creates graphs, maps and tables with the aim of contributing to the making of rational and timely decisions in the Cenaprece national VBD program.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdzul/densnv}
  \item \url{https://fdzul.github.io/densnv/}
  \item Report bugs at \url{https://github.com/fdzul/densnv/issues}
}

}
\author{
\strong{Maintainer}: Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com} (\href{https://orcid.org/0000-0002-2085-9141}{ORCID})

Other contributors:
\itemize{
  \item Fabian Correa-Morales \email{fabiancorrea@msn.com} (\href{https://orcid.org/0000-0002-6193-1242}{ORCID}) [contributor]
}

}
\keyword{internal}
