% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_eggs.R
\name{estimated_eggs}
\alias{estimated_eggs}
\title{Estimated cases of dengue and eggs}
\usage{
estimated_eggs(
  path_dengue,
  localidad,
  year,
  fac_axes,
  municipio,
  path_ovitraps
)
}
\arguments{
\item{path_dengue}{It is the path of the dataset of the sinave}

\item{localidad}{Its the locality target}

\item{year}{is the year}

\item{fac_axes}{is the factor for the dual axis}

\item{municipio}{is the municipality}

\item{path_ovitraps}{It is the path of the dataset of ovitrap}
}
\description{
Estimated cases of dengue and eggs
}
\examples{
1+1
}
