% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_heatmap.R
\name{mp_heatmap}
\alias{mp_heatmap}
\title{The mp_heatmap function generates a heat map.}
\usage{
mp_heatmap(
  locality,
  cve_edo,
  geocoded_dataset,
  status_caso,
  week,
  kernel,
  alpha,
  zoom,
  map_type,
  static,
  palette = NULL
)
}
\arguments{
\item{locality}{is the locality locality target.}

\item{cve_edo}{is a string of the state id.}

\item{geocoded_dataset}{is geocoded dataset.}

\item{status_caso}{It is a numerical value to select the Status of the case. There are three options, 1 for the probable cases database, 2 for the confirmed cases database, and 3 for the discarded cases database.}

\item{week}{It is epidemiological week.}

\item{kernel}{It is kernel density for select the blocks.}

\item{alpha}{is a numerical parameter that controls the transparency of the heatmap. Values range from 0 to 1, where 0 is completely transparent and 1 does not make the heat map transparent.}

\item{zoom}{map zoom; an integer from 3 (continent) to 21 (building), default value 10 (city).}

\item{map_type}{character string providing google map theme. options available are "terrain", "satellite", "roadmap", and "hybrid"}

\item{static}{is a logical valur, if static == TRUE the map is static, else (statis = FALSE) the maps is interactive.}

\item{palette}{is the palette for kde. example palette = viridis::turbo.}
}
\value{
ggplot object
}
\description{
The mp_heatmap function generates a heat map.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
