% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_serotype.R
\name{mp_serotype}
\alias{mp_serotype}
\title{This function creates maps of serotypes by municipality.}
\usage{
mp_serotype(path_sinave, cve_edo, palette)
}
\arguments{
\item{path_sinave}{is the directory where the files are located.}

\item{cve_edo}{is the id of state.}

\item{palette}{is is the color palette.}
}
\value{
a mapview objetc
}
\description{
This function creates maps of serotypes by municipality.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
