% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_treemap.R
\name{mp_treemap}
\alias{mp_treemap}
\title{Treemap of dengue cases}
\usage{
mp_treemap(snv_dataset, country, cve_edo = NULL)
}
\arguments{
\item{snv_dataset}{is the sinave dataset}

\item{country}{It is a logical value to indicate whether the data of the SINAVE is from the country (TRUE) or from a specific state (FALSE)}

\item{cve_edo}{is the state id of INEGI.}
}
\description{
Treemap of dengue cases
}
