% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ cases_week.R
\name{name is the name}
\alias{name is the name}
\title{Cases of dengue per week.}
\arguments{
\item{path}{is the path of the sinave files}

\item{mpo}{is the name of municipality}

\item{year}{is the year the dataset}
}
\value{
a ggplot2 object
}
\description{
Cases of dengue per week.
}
\examples{
1+1
}
