% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_geocode.R
\name{save_geocode}
\alias{save_geocode}
\title{save_geocode}
\usage{
save_geocode(geocode, data, directory, infile)
}
\arguments{
\item{geocode}{It is the geocoded dataset.}

\item{data}{is the dataset}

\item{directory}{This is the folder where the dataset will be saved.}

\item{infile}{is the name of the dataset file}
}
\value{
a RData object
}
\description{
save_geocode
}
\examples{
1+1
}
