% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_read_vbd.R
\name{sin_read_vbd}
\alias{sin_read_vbd}
\title{sin_read_vbd}
\usage{
sin_read_vbd(x, vbd, arbovirus)
}
\arguments{
\item{x}{is the directory where the files are located.}

\item{vbd}{is the parameter for define the vector-borne diseases.}

\item{arbovirus}{is a logical value for define the group of vector-borne diseases, if is TRUE, the define DENV, CHIKV & ZIKV, else for other etvs.}
}
\value{
a dataframe
}
\description{
The sin_read_vbd function reads the dataset of sinave vector-borne diseases.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
