% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_heatmap_edo.R
\name{mp_heatmap_edo}
\alias{mp_heatmap_edo}
\title{The mp_heatmap_edo function generates a heatmap by state.}
\usage{
mp_heatmap_edo(cve_edo, geocoded_dataset, week, kernel, alpha, palette)
}
\arguments{
\item{cve_edo}{is a string of the state id.}

\item{geocoded_dataset}{is geocoded dataset.}

\item{week}{It is epidemiological week.}

\item{kernel}{It is kernel density for select the blocks.}

\item{alpha}{is a numerical parameter that controls the transparency of the heatmap. Values range from 0 to 1, where 0 is completely transparent and 1 does not make the heat map transparent.}

\item{palette}{is the palette for kde. example palette = viridis::turbo.}
}
\value{
ggplot object
}
\description{
The mp_heatmap_edo function generates a heatmap by state.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
