% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{read}
\usage{
read(path, vbd, complete)
}
\arguments{
\item{path}{is the directory where the files are located.}

\item{vbd}{is the parameter for define the vector-borne diseases.}

\item{complete}{is a logical value for define the group of vector-borne diseases, if is TRUE, the define DENV, CHIKV & ZIKV, else for other etvs.}
}
\value{
a dataframe
}
\description{
The read function reads the dataset of sinave vector-borne diseases.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
